refactor footer into modular components
